browseruse-with-agentrun-demo/requirements.txt
# AgentRun SDK（用于 Sandbox 管理）
agentrun-sdk[playwright,server]>=0.0.8

ag-ui-protocol>=0.1.10


# BrowserUse 和浏览器自动化
browser-use>=0.1.0

# 环境变量管理
python-dotenv>=1.0.0

# 重试机制
tenacity>=8.2.0

# HTTP 客户端（用于 Sandbox API 调用）
httpx>=0.25.0

# Web 服务器（VNC Server）
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# 数据模型
pydantic>=2.0.0

# 异步支持
aiofiles>=23.0.0

# 可选：日志增强
loguru>=0.7.0

# 可选：配置管理
pyyaml>=6.0

